move index-2 to index
