+++ src/pages/index.js
@@ -1,93 +1,165 @@
-import React, { useState } from "react"
-import PropTypes from "prop-types"
-import { graphql, useStaticQuery } from "gatsby"
-import { useBottomScrollListener } from "react-bottom-scroll-listener"
+import React, { useState, useEffect } from "react"
+import { graphql, useStaticQuery, navigate } from "gatsby"
+import _ from "lodash"
 
-import Layout from "../components/layout"
-import SEO from "../components/seo"
-import GridView from '../components/GridView'
-import ExcerptCard from '../components/ExcerptCard'
-import GalleryItem from '../components/GalleryItem'
+import Seo from "@components/seo"
+import FlexBox from "@components/basic/FlexBox"
+import Nav from "@components/common/Nav"
+import Banner from "@components/home/Banner"
+import Footer from "@components/footer"
+import IntroCard from "@components/home/IntroCard"
+import ListPreview from "@components/home/ListPreview"
+import PlayListPreview from "@components/home/PlayListPreview"
+import ResponsiveBlock from "@components/home/ResponsiveBlock"
 
-const IndexPage = ({ location, displayPerTime }) => {
+import { scrollToAnchor, assignLanguages } from "@src/helpers"
+import { devices, sizes, colors, about } from "@constants/home"
+
+const IndexPage = () => {
   const data = useStaticQuery(graphql`
-    {
-      posts: allMarkdownRemark(
-        filter: {frontmatter: {path: {regex: "/\/(blog|gallery)\/.*/"}}}
+    query IndexQuery {
+      site {
+        ...SiteMetaFragment
+      }
+
+      avatar: file(relativePath: {eq: "miksin-avatar.jpg"}) {
+        childImageSharp {
+          fluid {
+            src
+          }
+        }
+      }
+
+      blogs: allMarkdownRemark (
+        filter: {frontmatter: {path: {regex: "/\/blog\/.*/"}}}
         sort: { order: DESC, fields: [frontmatter___date] }
-        limit: 10
+        limit: 3
       ) {
+        ...ArticlePreviewFragment
+      }
+
+      gallery: allMarkdownRemark (
+        filter: {frontmatter: {path: {regex: "/\/gallery\/.*/"}}}
+        sort: { order: DESC, fields: [frontmatter___date] }
+        limit: 5
+      ) {
+        ...GalleryPreviewFragment
+      }
+
+      nonSenseImages: allFile(filter: {extension: {regex: "/(jpg)|(png)|(jpeg)/"}, relativeDirectory: {eq: "nonsense"}}) {
         edges {
           node {
-            frontmatter {
-              path
-              title
-              excerpt
-              fromNow: date(fromNow: true)
-              date: date(formatString: "YYYY-MM-DD")
-              featuredImage {
-                childImageSharp {
-                  fluid {
-                    src
-                    aspectRatio
-                  }
-                }
-              }
-            }
+            ...FeaturedImageFragment
           }
         }
       }
     }
   `)
 
-  // extend display length when detect scroll to bottom
-  const newDisplayNum = now => Math.min(data.posts.edges.length, now + displayPerTime)
-  const [display, setDisplay] = useState(newDisplayNum(0))
-  useBottomScrollListener(() => {
-    setDisplay(newDisplayNum(display))
-  })
+  const [vw, setVw] = useState(undefined)
+  const [vh, setVh] = useState(undefined)
+  const [lang, setLang] = useState('en')
+  useEffect(() => {
+    /* eslint-disable no-undef */
 
-  const posts = data.posts.edges.slice(0, display).map(edge => {
-    const { frontmatter } = edge.node
-    const category = frontmatter.path.match(/^\/(.*?)\//)[1]
-    const viewType = category === 'gallery' ? GalleryItem : ExcerptCard
-    let inverseRatio = category === 'blog' ? 0.5 : 0
-    if (frontmatter.featuredImage) {
-      inverseRatio += 1 / frontmatter.featuredImage.childImageSharp.fluid.aspectRatio
+    // set window height (prevent url bar problems in mobile)
+    if (window) {
+      setLang(assignLanguages())
+      setVw(window.innerWidth)
+      setVh(window.innerHeight)
     }
 
-    return {
-      id: frontmatter.path,
-      inverseRatio,
-      viewProps: {
-        frontmatter: edge.node.frontmatter,
-      },
-      viewType,
-    }
-  })
+    /* eslint-enable no-undef */
+  }, [])
+
+  const bannerStyles = {}
+  if (vh) bannerStyles.minHeight = vh
+
+  const {
+    title,
+    subtitles,
+    author,
+    links,
+    socialLinks,
+  } = data.site.siteMetadata
+
+  const blogArticles = data.blogs.edges.map(edge => edge.node.frontmatter)
+  const galleryArticles = data.gallery.edges.map(edge => edge.node.frontmatter)
+  const nonSenseImages = _.shuffle(data.nonSenseImages.edges.map(edge => edge.node.childImageSharp.fluid.src))
+  const blockTitleSize = vw && vw <= devices.mobile ? sizes.scrollTitleMobile : sizes.scrollTitle
 
   return (
-    <Layout location={location}>
-      <SEO title="Home" />
-      <div className="center-view">
-        <h2>Recents</h2>
-        <div className="latest-post-list">
-          <GridView
-            viewList={posts}
+    <>
+      <Seo title="Home" />
+      <Nav links={links}/>
+      <Banner
+        styles={bannerStyles}
+        onEntry={() => scrollToAnchor('scroll-block-1', sizes.nav)}
+        words={{
+          title: [title],
+          subtitle: subtitles,
+        }}
+      />
+      <FlexBox column center>
+        <ResponsiveBlock
+          id="scroll-block-1"
+          bgColors={colors.lightBlue}
+          title={{
+            show: true,
+            text: 'ABOUT',
+            color: colors.white,
+            size: blockTitleSize,
+          }}
+          onEntry={() => scrollToAnchor('scroll-block-2', sizes.nav)}
+          onTitle={() => navigate('/about')}
+        >
+          <IntroCard
+            avatar={data.avatar.childImageSharp.fluid.src}
+            contents={about[lang] || about.en}
+            socialLinks={socialLinks}
           />
-        </div>
-      </div>
-    </Layout>
+        </ResponsiveBlock>
+        <ResponsiveBlock
+          id="scroll-block-2"
+          bgColors={colors.white}
+          title={{
+            show: true,
+            text: 'Recent Posts',
+            color: colors.lightBlue,
+            size: blockTitleSize,
+          }}
+          onEntry={() => scrollToAnchor('scroll-block-3', sizes.nav)}
+          onTitle={() => navigate('/blog')}
+        >
+          <ListPreview articles={blogArticles.map((frontmatter, index) => ({
+            ...frontmatter,
+            thumbnailSrc: nonSenseImages[index % nonSenseImages.length],
+            thumbnailAlt: 'blog',
+          }))} />
+        </ResponsiveBlock>
+        <ResponsiveBlock
+          id="scroll-block-3"
+          bgColors={colors.lightBlue}
+          title={{
+            show: true,
+            text: 'Gallery',
+            color: colors.white,
+            size: blockTitleSize,
+          }}
+          onTitle={() => navigate('/gallery')}
+        >
+          <PlayListPreview articles={galleryArticles.map(frontmatter => ({
+            ...frontmatter,
+            src: frontmatter.featuredImage.childImageSharp.fluid.src,
+            aspectRatio: frontmatter.featuredImage.childImageSharp.fluid.aspectRatio,
+          }))} color={colors.white} />
+        </ResponsiveBlock>
+      </FlexBox>
+      <Footer
+        siteAuthor={author}
+      />
+    </>
   )
-}
-
-IndexPage.propTypes = {
-  location: PropTypes.object.isRequired,
-  displayPerTime: PropTypes.number
-}
-
-IndexPage.defaultProps = {
-  displayPerTime: 20
 }
 
 export default IndexPage
